add Debian stuff

+++ all_deps.sh
@@ -1,19 +1,30 @@
-packages='MariaDB-Manager-GREX MariaDB-Galera-server MariaDB-client rsync iproute net-tools grep findutils gawk'
+release_info=$(cat /etc/*-release)
+if [[ $(echo "$release_info" | grep 'Red Hat') != "" || $(echo "$release_info" | grep 'CentOS') ]]; then
+	packages='MariaDB-Manager-GREX MariaDB-Galera-server MariaDB-client rsync iproute net-tools grep findutils gawk'
 
-cd /home/ec2-user
-#rm -rf /home/ec2-user/packages
-rm -rf /home/ec2-user/repo
-#mkdir /home/ec2-user/packages
-mkdir /home/ec2-user/repo
+	cd /home/ec2-user
+	#rm -rf /home/ec2-user/packages
+	rm -rf /home/ec2-user/repo
+	#mkdir /home/ec2-user/packages
+	mkdir /home/ec2-user/repo
 
-for i in $packages
-do
-	list=`rpmdep $i 2> /dev/null | grep "depends upon" | sed "s/depends upon//" | sed "s/,/ /g"`
-	echo  $list
-	yum -y --downloadonly --downloaddir=/home/ec2-user/packages1/ reinstall $list
-	cp /home/ec2-user/packages1/* /home/ec2-user/packages/
-	yum -y --downloadonly --downloaddir=/home/ec2-user/packages1/ reinstall $i
-        cp /home/ec2-user/packages1/* /home/ec2-user/packages/
-done
+	for i in $packages
+	do
+	        list=`rpmdep $i 2> /dev/null | grep "depends upon" | sed "s/depends upon//" | sed "s/,/ /g"`
+        	echo  $list
+	        yum -y --downloadonly --downloaddir=/home/ec2-user/packages1/ reinstall $list
+        	cp /home/ec2-user/packages1/* /home/ec2-user/packages/
+	        yum -y --downloadonly --downloaddir=/home/ec2-user/packages1/ reinstall $i
+        	cp /home/ec2-user/packages1/* /home/ec2-user/packages/
+	done
 
+elif [[ $(echo "$release_info" | grep 'Ubuntu') != "" || $(echo "$release_info" | grep 'Debian') ]]; then
+
+	packages='mariadb-manager-grex mariadb-galera-server mariadb-client rsync iproute net-tools grep findutils gawk'
+	list_all_packages=`apt-rdepends $packages | sed "s/PreDepends://" | sed "s/Depends://" | sed "s/([^)]*)//g" | tr '\n' ' '`
+	list_all_packages=`echo $list_all_packages | sed "s/ awk/ gawk /g" | sed "s/debconf-2.0 /debconf /g" | sed "s/libstorable-perl /perl /g" | sed "s/perlapi-5.14.2 /perl-base /g" | sed "s/perl-dbdabi-94 /libdbi-perl /g"`
+	cd  /home/ec2-user/packages/
+	apt-get download $list_all_packages
+fi
 /home/ec2-user/create_repo.sh /home/ec2-user/repo /home/ec2-user/packages
+
